+++ src/Milestones/Model.php
@@ -7,6 +7,7 @@ class Model {
 
 	// Settings
 	protected $title;
+	protected $description;
 	protected $image;
 
 	/**
@@ -16,8 +17,9 @@ class Model {
 	 * @since 2.9.0
 	 **/
 	public function __construct( array $args ) {
-		isset( $args['title'] ) ? $this->title = $args['title'] : $this->title = __( 'Sample Milestone Title', 'give' );
-		isset( $args['image'] ) ? $this->image = $args['image'] : $this->image = '';
+		isset( $args['title'] ) ? $this->title             = $args['title'] : $this->title = __( 'Sample Milestone Title', 'give' );
+		isset( $args['description'] ) ? $this->description = $args['description'] : $this->description = __( 'This is a sample description.', 'give' );
+		isset( $args['image'] ) ? $this->image             = $args['image'] : $this->image = '';
 	}
 
 	/**
@@ -46,6 +48,17 @@ class Model {
 	}
 
 	/**
+	 * Get description for Milestone
+	 *
+	 * @return string
+	 * @since 2.9.0
+	 **/
+	protected function getDescription() {
+		return $this->description;
+	}
+
+	/**
+	 * Get template path for Milestone component template
 	 *
 	 * @return string
 	 * @since 2.9.0